fix bug groom feature & love story feature finished

+++ app/Http/Controllers/User/BrideManController.php
@@ -98,11 +98,17 @@ class BrideManController extends Controller
                 $filename = sprintf('%s_%s.%s', date('Y-m-d'), md5(microtime(true)), $file->extension());
                 $image_path = $file->move('storage/brideman', $filename);
             }else{
-                unlink($brideman->image);
+                if(file_exists($brideman->image)){
+                    unlink($brideman->image);
 
-                $file = $request->file('image');
-                $filename = sprintf('%s_%s.%s', date('Y-m-d'), md5(microtime(true)), $file->extension());
-                $image_path = $file->move('storage/brideman', $filename);
+                    $file = $request->file('image');
+                    $filename = sprintf('%s_%s.%s', date('Y-m-d'), md5(microtime(true)), $file->extension());
+                    $image_path = $file->move('storage/brideman', $filename);
+                }else{
+                    $file = $request->file('image');
+                    $filename = sprintf('%s_%s.%s', date('Y-m-d'), md5(microtime(true)), $file->extension());
+                    $image_path = $file->move('storage/brideman', $filename);
+                }
             }
         }else{
             $image_path = $brideman->image;
